# backend/api/openapi/swagger.yaml
definitions:
...
  sico-backend_internal_transport_http_dto_skill.CreateSkillData:
    properties:
      skill:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.Skill'
    type: object
  sico-backend_internal_transport_http_dto_skill.CreateSkillRequest:
    properties:
      agentId:
        type: string
      assetId:
        type: integer
      projectId:
        type: integer
    required:
    - assetId
    type: object
  sico-backend_internal_transport_http_dto_skill.CreateSkillResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.CreateSkillData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.DeleteSkillResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.GetSkillData:
    properties:
      skill:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.Skill'
    type: object
  sico-backend_internal_transport_http_dto_skill.GetSkillDetailsData:
    properties:
      files:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.SkillFile'
        type: array
    type: object
  sico-backend_internal_transport_http_dto_skill.GetSkillDetailsResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.GetSkillDetailsData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.GetSkillResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.GetSkillData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.ListSkillData:
    properties:
      hasNext:
        type: boolean
      skills:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.Skill'
        type: array
      total:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_skill.ListSkillResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.ListSkillData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.Skill:
    properties:
      agentId:
        type: string
      assetId:
        type: integer
      createdAt:
        type: integer
      creatorUsername:
        type: string
      description:
        type: string
      failReason:
        type: string
      id:
        type: integer
      name:
        type: string
      projectId:
        type: integer
      status:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.SkillStatus'
      updatedAt:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_skill.SkillFile:
    properties:
      content:
        type: string
      path:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_skill.SkillStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      SkillStatus_SKILL_STATUS_FAILED: '@gotag: json:"SKILL_STATUS_FAILED"'
      SkillStatus_SKILL_STATUS_UNKNOWN: '@gotag: json:"SKILL_STATUS_UNKNOWN"'
      SkillStatus_SKILL_STATUS_UPLOADED: '@gotag: json:"SKILL_STATUS_UPLOADED"'
      SkillStatus_SKILL_STATUS_UPLOADING: '@gotag: json:"SKILL_STATUS_UPLOADING"'
    x-enum-descriptions:
    - '@gotag: json:"SKILL_STATUS_UNKNOWN"'
    - '@gotag: json:"SKILL_STATUS_UPLOADING"'
    - '@gotag: json:"SKILL_STATUS_UPLOADED"'
    - '@gotag: json:"SKILL_STATUS_FAILED"'
    x-enum-varnames:
    - SkillStatus_SKILL_STATUS_UNKNOWN
    - SkillStatus_SKILL_STATUS_UPLOADING
    - SkillStatus_SKILL_STATUS_UPLOADED
    - SkillStatus_SKILL_STATUS_FAILED
  sico-backend_internal_transport_http_dto_skill.UpdateSkillData:
    properties:
      skill:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.Skill'
    type: object
  sico-backend_internal_transport_http_dto_skill.UpdateSkillRequest:
    properties:
      assetId:
        type: integer
      id:
        type: integer
    required:
    - assetId
    - id
    type: object
  sico-backend_internal_transport_http_dto_skill.UpdateSkillResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.UpdateSkillData'
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Sico AI Agent Platform API
  title: Sico API
  version: "1.0"
paths:
...
  /api/sico/skills:
    delete:
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.DeleteSkillResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
    get:
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.GetSkillResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Skill
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.CreateSkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.CreateSkillResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Skill
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.UpdateSkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.UpdateSkillResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
  /api/sico/skills/details:
    get:
      parameters:
      - in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.GetSkillDetailsResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
  /api/sico/skills/list:
    get:
      parameters:
      - in: query
        name: agentId
        type: string
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: projectId
        type: integer
      - enum:
        - 0
        - 1
        - 2
        - 3
        format: int32
        in: query
        name: status
        type: integer
        x-enum-comments:
          SkillStatus_SKILL_STATUS_FAILED: '@gotag: json:"SKILL_STATUS_FAILED"'
          SkillStatus_SKILL_STATUS_UNKNOWN: '@gotag: json:"SKILL_STATUS_UNKNOWN"'
          SkillStatus_SKILL_STATUS_UPLOADED: '@gotag: json:"SKILL_STATUS_UPLOADED"'
          SkillStatus_SKILL_STATUS_UPLOADING: '@gotag: json:"SKILL_STATUS_UPLOADING"'
        x-enum-descriptions:
        - '@gotag: json:"SKILL_STATUS_UNKNOWN"'
        - '@gotag: json:"SKILL_STATUS_UPLOADING"'
        - '@gotag: json:"SKILL_STATUS_UPLOADED"'
        - '@gotag: json:"SKILL_STATUS_FAILED"'
        x-enum-varnames:
        - SkillStatus_SKILL_STATUS_UNKNOWN
        - SkillStatus_SKILL_STATUS_UPLOADING
        - SkillStatus_SKILL_STATUS_UPLOADED
        - SkillStatus_SKILL_STATUS_FAILED
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_skill.ListSkillResponse'
      security:
      - BearerAuth: []
      tags:
      - skills
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"